Fixed errors in output printing

+++ rankmaniac-lib/rankmaniac-students/data/pagerank_reduce.py
@@ -12,8 +12,8 @@ for line in sys.stdin:
     values = line.split("\t")
     node_idx = int(values[0])
     
-    sum += int(values[1])
+    sum += float(values[1])
 new_rank = (0.85 * sum) + 0.15
 sys.stdout.write("%d\t%f" % (node_idx, new_rank))
-    
+#
